backend/requirements.txt
requests
colorlog
python-dotenv
Flask
Flask-Cors
oauthlib
requests_oauthlib
paramiko

--extra-index-url https://mkmartifactory.amd.com/artifactory/api/pypi/hw-orc3pypi-prod-local/simple
--trusted-host mkmartifactory.amd.com
at-scale-python-api